+++ lib/clientOptions.ts
@@ -1,0 +1,9 @@
+export interface ClientOptions {
+  apiHostname?: string;
+  defaultVersion?: string;
+}
+
+export const ClientOptionsDefaults: ClientOptions = {
+  apiHostname: "https://app.procore.com",
+  defaultVersion: "v1.0"
+};