+++ lib/refresh.ts
@@ -1,5 +1,5 @@
 import 'isomorphic-fetch'
-import _hostname from './hostname'
+import { ClientOptions, ClientOptionsDefaults} from './clientOptions'
 
 export interface RefreshConfig {
   id: string;
@@ -9,9 +9,10 @@ export interface RefreshConfig {
   refresh: string;
 }
 
-function refresh({ id, secret, uri, token, refresh }: RefreshConfig, hostname: string = _hostname) {
+function refresh({ id, secret, uri, token, refresh }: RefreshConfig, options: ClientOptions = ClientOptionsDefaults): Promise<any> {
+  const _options = Object.assign({}, ClientOptionsDefaults, options);
   return fetch(
-      `${hostname}/oauth/token?grant_type=refresh_token&$client_id=${id}&client_secret=${secret}&redirect_uri=${uri}&refresh_token=${refresh}`,
+      `${_options.apiHostname}/oauth/token?grant_type=refresh_token&$client_id=${id}&client_secret=${secret}&redirect_uri=${uri}&refresh_token=${refresh}`,
       { method: 'POST', headers: { 'Authorization': `Bearer ${token}` } }
     )
     .then(res => res.json())