+++ test/implicit.spec.ts
@@ -1,5 +1,5 @@
 import { expect } from 'chai'
-import { implicit } from './../lib'
+import { implicit } from '../lib/index'
 
 describe('implicit grant', () => {
   it('returns valid oauth url', () => {
@@ -7,6 +7,6 @@ describe('implicit grant', () => {
     const uri = 'example.com/redirect';
     expect(implicit({ id, uri })).to.equal(
       `https://app.procore.com/oauth/authorize?response_type=token&client_id=${id}&redirect_uri=${uri}`
-     )
+    )
   })
 })