+++ test/client.spec.ts
@@ -8,7 +8,7 @@ const rfi = { id: 1, subject: 'Create RFI Subject', assignee_id: 2945 };
 const idsToDelete = [{ id: 1 }, { id: 2 }];
 const token = 'token';
 const hostname = 'https://app.procore.com';
-const headers = { 'Accept': 'application/json', 'Content-Type': 'application/json', 'X-Procore-Sdk-Version': sdkVersionHeader };
+const headers = { 'Accept': 'application/json', 'Content-Type': 'application/json', 'Procore-Sdk-Version': sdkVersionHeader };
 
 describe('client', () => {
   it('uses a custom formatter', async () => {