+++ lib/refresh.ts
@@ -1,12 +1,32 @@
 import 'isomorphic-fetch'
-import _hostname from './hostname'
+import { ClientOptions, convert } from './clientOptions'
 
-function refresh({ id, secret, uri, token, refresh }, hostname = _hostname) {
-  return fetch(
-      `${hostname}/oauth/token?grant_type=refresh_token&$client_id=${id}&client_secret=${secret}&redirect_uri=${uri}&refresh_token=${refresh}`,
-      { method: 'POST', headers: { 'Authorization': `Bearer ${token}` } }
-    )
-    .then(res => res.json())
+export interface RefreshConfig {
+  id: string;
+  secret: string;
+  uri: string;
+  token?: string;
+  refresh: string;
+}
+
+async function refresh({ id, secret, uri, token, refresh }: RefreshConfig, options: ClientOptions | string): Promise<any> {
+  const res = await fetch(
+    `${convert(options).apiHostname}/oauth/token`,
+    {
+      method: 'POST',
+      headers: {
+        'Content-Type': 'application/json'
+      },
+      body: JSON.stringify({
+        'grant_type': 'refresh_token',
+        'client_id': id,
+        'client_secret': secret,
+        'redirect_uri': uri,
+        'refresh_token': refresh
+      })
+    }
+  );
+  return res.json();
 }
 
 export default refresh