+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 * Remove `@procore/js-sdk-endpoints` dependency
 * Add support for revoking token `revoke()`
 * Add `Procore-Sdk-Version` header to all requests
+* Add `Procore-Sdk-Language` header to all requests
 * Updated `destroy` method to `delete` method
 
 ### Breaking Changes